move paths.js to middleware util folder
