+++ src/middleware/data.js
@@ -16,7 +16,7 @@ limitations under the License.
 import rejectingFetch from './utils/rejectingFetch';
 import invariant from 'invariant';
 import { headersGet, headersPost, headersPut, headersDelete } from './utils/headers';
-import { dataApiPath } from './paths';
+import { dataApiPath } from './utils/paths';
 import { noteSave, noteFailure } from '../store/saveState';
 
 /******
