+++ src/middleware/genbank.js
@@ -16,7 +16,7 @@
 import rejectingFetch from './utils/rejectingFetch';
 import invariant from 'invariant';
 import { headersGet, headersPost, headersPut, headersDelete } from './utils/headers';
-import { extensionApiPath } from './paths';
+import { extensionApiPath } from './utils/paths';
 import uploadFiles from './utils/uploadFiles';
 
 const extensionKey = 'genbank';
