+++ src/middleware/projectFile.js
@@ -16,7 +16,7 @@
 import rejectingFetch from './utils/rejectingFetch';
 import invariant from 'invariant';
 import { headersGet, headersPost, headersDelete } from './utils/headers';
-import { projectFilePath } from './paths';
+import { projectFilePath } from './utils/paths';
 
 const contentTypeTextHeader = { headers: { 'Content-Type': 'text/plain' } };
