+++ src/middleware/utils/paths.js
