+++ test/middleware/report.js
@@ -17,7 +17,7 @@
 import { expect } from 'chai';
 import rejectingFetch from '../../src/middleware/utils/rejectingFetch';
 import { headersGet } from '../../src/middleware/utils/headers';
-import { reportApiPath } from '../../src/middleware/paths';
+import { reportApiPath } from '../../src/middleware/utils/paths';
 
 describe('Middleware', () => {
   describe('Reporting', () => {
