+++ test/server/data/REST/fetch.spec.js
@@ -1,7 +1,7 @@
 import { expect } from 'chai';
 import fetch from 'isomorphic-fetch';
 import Project from '../../../../src/models/Project';
-import { dataApiPath } from '../../../../src/middleware/paths';
+import { dataApiPath } from '../../../../src/middleware/utils/paths';
 
 //noinspection ES6UnusedImports
 import devServer from '../../../../server/server'; // starts the server which will be accessed by methods below
